Doc strings now use the following format for parameters:
Description of method.

Parameters
----------
param1 : <type>(, optional/method specific)
	<description> If the <type> accepts one of a list of strings,
	list them in the doc string as 'Options are x, y, and z'.
	Units: <if applicable>
	Default: 'x'

Returns
-------
return1 : <shape> <type>
	Description of return. if the size is 1, shape is not required.

method specific parameters are a subset of optional parameters,
exclusively for Conductivity object parameters where some parameters won't
be necessary for all methods.

Return formats are still in progress. Additionally, hyperlinks are in the
works so that the links can be added to move users from docstrings to the
appropriate theory passages. Currently, the docstrings write out the "Defaults"
statement but this will be fixed in a future push.